Api/requirements.txt
asgiref>=3.6.0<4
Django==4.2.7
pytz==2021.1
sqlparse==0.4.1
sqlalchemy==2.0.9
djangorestframework==3.14.0
djangorestframework-simplejwt
django-cors-headers==3.5.0
mysqlclient>=2.0
Pillow==10.1.0


# pika==1.1.0